Passes: Create MIPS arithmetic pass
Create languages/passes for converting a Tiger IR program into a
program with MIPS arithmetic instructions (replacing TigerIR's
assign, add, sub, div, mult, and, or with MIPS equivalents).

We're still deciding whether we should use Nanocaml for this project.
This creates a quick demo of using Nanocaml and helps us get a better
idea of the pitfalls involved. As part of this, I discovered a few
bugs in Nanocaml:
- No support for normal variants as arguments to poly variants (bug in checker)
- No support for non-rec pass functions
- Using open inside of a language is an error

As well as some limitations:
- Can't use | patterns
- Can't use guard patterns
- Can't skip defining a pass for a type that doesn't change (or do function x -> x)
- Can't get rid of an entire type in a language

I'm still interested in using Nanocaml but I'm becoming more & more open
to the idea of just writing these passes manually, as there's a very similar
amount of boilerplate if things continue as they've gone so far.