Passes: Remove unnecessary pass for "op"
I previously defined "op" as part of the NestedIr language, which
requires it to stick around in every subsequent language (even if
it's no longer used). That would mean having to handle a pass for it
in every subsequent language. Instead, we can just move it out of
the language definition and everything works as expected!