Passes: Add conditional/branching pass
This pass translates TigerIR branching instructions & conditionals
into their MIPS forms. The main two challenges here are making sure
that we convert `op` parameters into virtual registers (which is
accomplished using the `use_op` function) and handling comparisons
to zero rather than relative comparisons between registers. This
second goal is accomplished by generating a sub instruction before
every instruction that compares to zero (note: there's room for
optimization here in the case of comparisons to immediates or 0).
To do this, we generate a `Sub` instruction that writes to a virtual
register. During register allocation, we can hopefully eliminate
the extra register here in the majority of cases.