Passes: Implement array operation translation
This pass translates operations like array assign, store, and load
to MIPS assembly. Some weirdness: ArrayAssign creates a block of
assignments to each value in the array rather than a loop. This also
enables lw/sw instructions for use in the next stages (register
allocation, function calls) for access to the stack/spilling.