Testing: implement to_string for output, enable flattening, test program
This sets up a simple pipeline that takes the input program in TigerIR
and converts every single pass to get a final program. It then flattens
the instruction sequence into a linear program and prints it out. A
simple test program using a variety of instructions is included in the
Codegen module, and the result of compiling that program is printed to
the command-line when the compiler is run. Be on the lookout for __'s,
which indicate holes for register allocation to fill in.