Passes: Get rid of `Obj.magic` in `flatten`
Previously, when flattening a MipsLiStack.instr list, we needed to use
Obj.magic to perform a downcast to MipsFlat.instr list since the compiler
did not recognize the lack of a `Block instr being possible (the case was
previously handled). Since the MipsFlat.instr type is a strict subset of
MipsLiStack.instr, we actually have another solution: OCaml allows us to
match against an entire variant type using the #t syntax. So instead of
performing the cast using Obj.magic, we can have the match case "assert"
that the matched value is already a valid MipsFlat.instr by doing:

| (#MipsFlat.instr as instr)::rest -> ...

Pretty nifty!