Alloc: Implement collect_vars function for register allocation
This function traverses a tree of instructions and collects every
variable/virtual register present. Since it's designed as both an
instruction-level and a block-level operation, this can come in
handy for future register allocation analyses (we just have to be
careful when dealing with nested blocks). Specifically, this can
help us get an idea of live ranges because it'll identify every
single var used in an instr or range of instrs.