Begin fleshing out stack frame details
This starts specifying decisions about how the stack frame should
be laid out, as well as providing some code to generate function
prologs and epilogs. We're still not quite at the point where
register allocation is ready to happen, so this needs a little
bit more work (namely, figure out what the spill positions for each
variable should be; what registers are available for allocation;
do we want to follow MIPS conventions even if they're not directly
needed -- e.g. aligning $sp to 8?).