Codegen: Hook up prolog/epilog generation
The Passes.compile function now takes a function rather than a list
of instructions. This change is step 1 in the process of performing
full compilation of a program, and will enable us to do things like
apply register allocation results to a function's parameters.

Right now, this change only really does one thing: it takes the
function's data segment and uses that to generate the stack building
and tear-down code for the function.