Register allocation: Implement naive allocator algorithm
Naive allocation is super simple: instead of allocating physical
registers for each virtual register, we simply spill them all to
the stack. When we need to use the virtual register, we quickly
load/store it before/after its use. Some work may need to be done
in the case of arrays here... Also, note that this isn't hooked
up to anything yet.