Register Allocation: Implement pass to apply allocations
This was unexpectedly rather challenging and the code has a long
way to go before being fast or clean, but this code reflects a
crucial milestone in register allocation. Specifically, this code
performs the transformation to an instruction list that applies
register assignments. If a register is spilled, it grabs it off
the stack and uses it (or calculates its value and stores it onto
the stack). Otherwise, it simply uses the register as is. We only
allow two registers for spilled values: $at and $v1 (since neither
of these need to be used anywhere else). $at is used for both the
result and first argument to any instruction, while $v1 is used
as the second argument to the instruction. This is always safe,
as no arguments can ever be in/out and spilling only gets applied
at the instruction level, meaning that the code never assumes
that $at/$v1 aren't clobbered.

Quick note: when building a function's stack, we need to know the
spilled amount in order to accurately increment the stack pointer.
Some part of the allocator should return this value, so that the
compiler is able to take it into account when building the stack.