Alloc: Use MipsFlat IR instead of MipsMem
Change all of the allocator to operate on the MipsFlat IR rather than MipsMem.
This is a somewhat weird choice because at MipsFlat we already have a mix of
virtual and physical registers, but MipsFlat deals with register demand much
better than MipsMem. This is because of how function calls require only a
maximum of 4 physical registers, but the TigerIR instruction allows for many
more registers. In order to minimize the demand for registers during the spill,
we prefer to avoid the `call` and `callr` instructions here.