Alloc: Recognize register names when allocating virtual registers
In order to perform the change from MipsMem to MipsFlat in the allocator,
we need to recognize physical register names when they're present. For
example, if we're already mapping a function argument to a0, we don't want
to treat a0 as a virtual register and map it to another physical register.
Instead, we need to add a 1:1 mapping and make sure that the naive allocator
doesn't ever override an existing mapping.