Alloc: Return amount of new stack space that must be allocated
Because of the move from MipsMem to MipsFlat IR, we've designed the
allocator to occur before we perform the stack building for a procedure.
As a result, we no longer have access to the locals_base/temps_base
offsets from the stack builder -- instead, we need to calculate these in
the register allocator. This isn't optimal and I'd like to revisit it,
but I need to think about architecture a little bit. Note that the stack
building depends on the allocator as well, since the allocator can demand
more space on the stack! This takes the form of the new_spills variable,
which is now returned from the allocator along with the mapping and
reflects the amount of extra space needed on the stack for allocator
temps.