Alloc: Fix stack temps to start _AFTER_ the locals and count by words
This is a quick fix that takes care of a bug in the allocator. Previously,
when allocating stack space for a temp, we could place it at offset 0 from
the end of the stack locals. Because memory grows downwards, this would
really overlap with the last value on the stack! Thus, we could end up with
a temp overwriting a local variable.

Additionally, we were previously counting stack temps by 1's (AKA bytes),
but each stack temp will always be 1 word. The quick fix here is just
counting by 4's instead of by 1's.