MipsMem pass: Perform array indexing by word instead of by byte
This makes sure to multiply by 4 (word size) when performing array
indexing: in the code generation for arrays, we need to keep in mind
that every cell (assuming it's an int/float and we don't have 2d
arrays) is 1 word in size. As a result, we have to translate user-
provided indices (multiples of 1 byte) into memory indices (multiples
of 4 bytes/1 word) by performing a multiplication. In order to
optimize the outputted code, I use the `sll` immediate left-shift
instruction, but this requires a few changes in the rest of the
compiler to handle the new `sll` instruction. When the index is
constant, I simply multiply it by 4 before generating the final
instruction.