MipsCond pass: Use temporary comparison register for compare-to-z breaks
Quick fix in b{g,l}{t,e}z instructions: use a new virtual register rather
than `x` as the comparison register. Since we don't want to modify the
operand registers in a comparison, we have to make sure that we're
placing the subtraction's result in a new virtual register and comparing
that value to 0 in the actual MIPS instruction.