Alloc: Recognize function parameters as locals
This essentially applies the optimization for locals to parameters (if
it's a parameter or local, it already has some space allocated so we
should use that instead of creating new space on the stack), though it's
a little more complex because some params are already assigned a physical
register.

If we pass a parameter into the function, there's 2 cases. If it's one of
the first 4 params, it's already assigned to one of $a0..$a3, so we don't
want to allocate a new register for it and we don't need to spill either.
If it's one of the other parameters, it's stored on the stack right before
$fp. In this case, we need to figure out whether a virtual register
corresponds to one of these parameters and if so, we either pick the
register it's in or spill it to the parameter portion of the stack.

of the first