Alloc: Handle spilling for local arrays
When using virtual registers for an array, the virtual register should
be equal to the address of the start of the array (AKA &a[0] instead
of a[0]), since we dereference $arr_register+offset. Whenever we have
a local array, we need to make sure that what we put in the register
is &a[0] (or a + 0). If it's not a local array, we can assume it's a
function parameter (in which case it's fine to just load that from
the stack because it's a pointer to the start of the array) but for
local arrays we need to add the stack offset and $fp without any load
operation.

To do this, we add the SpillArr (for "spill array") allocation, which
is reserved for local arrays that get spilled to the stack. This
simply means that the allocator should perform an add instead of a lw
when performing the spill's reads/writes. Due to this change, we had
to reorganize much of the `apply_alloc_*` code to handle this new
edge case.