Driver: Hook up allocator to the main compiler
This is the main part of enabling the naive register allocator. Essentially,
we're implementing the interface between the instruction selection passes
and the register allocation code. Due to these being isolated in two modules
and the allocator module already depending on the instruction selection
module, we've designed the interface to avoid adding new dependencies. The
Passes.compile function now takes a ~allocator parameter that provides the
two steps of allocation: the naive allocator + the apply_allocations
function.

A bit of restructuring is also done in the Passes.compile function to break
up these two stages. Specifically, it performs all the main passes, then
allocates, and finally builds the stack using the results from the allocator.

The stack-building code also had to be modified here to take the ~new_spills
parameter, which tells it how much extra space to allocate on the stack for
spilled variables.

Finally, one tiny change is adding a label to the top of each function. This
will eventually allow us to string together a program composed of multiple
compiled functions, since each function name can now get resolved to a label.