Prepend physical register names with $ in IR
We need to have some way of knowing which registers are virtual vs. physical,
in case a user names a virtual register e.g. "a0" or "zero". The easy solution
here is simply prepending the register names with $ wherever they occur, and
making sure the parser doesn't allow for the names with a $ prepended.