ISel: Add `.set noat` directive to allow use of $at register
Right now, the $at register is used by the naive allocator for spilled
temporaries. This was chosen since it's one of the few registers that
shouldn't interfere with anything else (assuming we don't use any
instructions that require assembler temporaries). Other choices here
are $v1 (which is already used) and $k0/$k1/$gp -- all of which have
important uses. By default SPIM doesn't allow us to use the $at
register but by adding this directive it relaxes this restriction.