Alloc: fix edge-case when spilling 1-cell local array
When detecting local arrays during spilling (see previous commits that
worked on this), I used the size of the local to determine whether to
treat it as an array or not. However, this is inaccurate for 1-cell
arrays, since their size will still be 4. To get around this, I simply
keep a bool indicating whether each local was an array or not and then
use that instead of checking sizes.