Alloc: immediately add param registers to mapping
While we're fine with the naive allocator handling param registers,
the Chaitin-Briggs allocator will need those to be preallocated,
since it would attempt to add them to the dataflow analysis if they
weren't known to be pre-allocated (actually, the situation here is
a little more complicated since we check for $'s in register names
in the Chaitin-Briggs allocator, but we still want to make sure it
ignores these registers when building the RIG and it doesn't hurt
to filter them out early on). This also considerably simplifies the
various edge cases handled by the naive allocator, so that's a big
plus in my book!