Alloc: let allocators take a mapping of pre-allocations to build on
Since we want to chain together the Chaitin-Briggs and naive allocators
(Chaitin-Briggs will use the naive allocator to handle spills), it makes
sense to modularize allocators a little bit. Here, we make allocators
take in a pre-mapping, which contains allocations that were already
determined (and the new allocator will know to ignore these). To demo
this a little bit, we now have a function for Chaitin-Briggs in the Alloc
module that forwards to the naive allocator (but first does liveness
analysis using the Dataflow module!). The default allocator is now
Chaitin-Briggs, so we'll have an easier time seeing what happens as the
new allocator progresses.