Dataflow: Perform liveness analysis in reverse toplogical order
Since we're performing liveness analysis as a backwards dataflow, it
doesn't make much sense to solve the dataflow equations forwards (as we
were doing in the middle end). While I'm not sure it impacts the final
result, we can get a significant performance improvement by performing
the dataflow backwards. To do this, we can use the reverse topological
ordering of the graph and fold our map over this. Since there's no
reverse_fold on the topo-sort included in OCamlgraph, we can just topo-
sort the reverse graph g^R (which we can get from the mirror function
in the Oper module). Side note, I'm not actually sure that this is
equal to the reverse topological sort of the graph but I think it should
still help with performance.

Source: https://www.cs.colostate.edu/~mstrout/CS553/slides/lecture03.pdf