Chaitin-Briggs: Create RIG module
The RIG is the register interference graph, which contains a vertex
for every variable and an edge between any two variables that are
simultaneously live. Right now, this just defines the module for
the graph used in the RIG, but eventually this will also take care
of operations such as building, coloring, etc.