Chaitin-Briggs: Build RIG and color it
This implements the code to build the RIG, which works like so: first
we add all variables in the CFG as vertices. Then, we iterate through
every in_set/out_set and take every 2-permutation of variables to add
as an edge: e.g. if {a, b, c} is the in-set, we need edges a-b, a-c,
and b-c.

Once we have the RIG built, we can just call the graph coloring code
incldued in OCamlgraph to do the rest for us. This takes a number that
represents the number of colors to use (& spills the rest), so we pass
in the # of registers available for allocation. However, this may lead
to a sub-optimal coloring, since we have no input on which vertices
should be spilled. It might be better to pass the # of variables in (so
that none are spilled) and then manually pick which ones to keep based
on how often each variable is used.