RIG: ignore pre-allocated values while building RIG
Previously, the RIG didn't check to see if a value was pre-allocated
during the build phase. This is an issue when dealing with the first
4 params of a function, since these will always be pre-allocated and
we don't want to allocate them to 2 separate registers. Instead, we
avoid adding vertices or edges if the variables have already been
assigned to a register (based on the passed in pre-allocation table).