ISel: Add a pass to load params from the stack
Before register allocation, we want to load params from the stack into
virtual registers. This is useful because (1) it actually defines those
params somewhere so that we can use them and (2) it resolves the tricky
situation where we need to figure out somewhere to load params after
performing Chaitin-Briggs. Since they're assigned to virtual registers
here, the allocator knows that these registers are live from the get-go.
This isn't entirely optimal because it can introduce some deadcode and
is a bit too conservative (we can get around this by copying between
registers and inserting the loads later), but it's a simple solution
that's very easy to get right in 100% of cases. Note that this really
only needs to happen before Chaitin-Briggs, since the naive allocator
deals with this by always loading before using every variable (and it
loads from the param's position on the stack).