Codegen/Examples: Added input, output, and allocator command line arguments and examples
Added the -i (input filename), -o (output filename), and --allocator (allocator type) options
when executing the backend through the terminal. The input file is required and specifies the
TigerIR file to be used as input, while the output file, also required, specifies where to
output the assembly. The allocator is optional, and defaults to a naive allocator. You can
choose between the inputs "Naive" (naive allocator) or "Graph" (Chaitlin-Briggs Graph Allocator).

The examples directory has example TigerIR files to be used as inputs to the command.