ISel: Save/restore all allocated registers
In order to safely allocate $t0..$t9, $s0..$s7, and $a0..$a3, we
need to make sure to save these registers on function boundaries.
By convention, $s0..$s7 are saved by the callee and the rest are
saved by the caller. This isn't the prettiest (or most efficient)
thing as it's implemented right now, but it's good enough to work.
I might revisit this in the future though and try to make it a
little more efficient (by only saving registers that NEED to be
saved.