ISel: rename labels to be unique in their function
In TigerIR, functions introduce a private scope for labels. Labels in
MIPS assembly, however, are all global. This means that two names can
clash if the same label is used in multiple functions. To get around
this, we can simply prepend the label name with the name of the
function.