ISel: Don't overwrite function result in callr restore
When restoring caller-saved registers in the callr instruction, it
overwrites all $t* and $a* registers. If the result of a function
call is stored in one of these registers, it will get overwritten if
we're not careful. To make sure it doesn't get overwritten by the
register restoring, we just have to make sure to copy $v0 to the
assigned register __after__ the restore occurs.