Chaitin-Briggs: ignore local arrays when building RIG
This is very similar to the local arrays issue we encountered in the
naive allocator. In this case, the easiest approach to get it working
quickly is just to yield to the naive allocator (since it already
does this), but this probably isn't the optimal way to do it -- we
could try something similar to the stack parameter loading we did.
Will definitely need to revisit this decision in the future!