Rig: create edges between currently defined variable and members of out set
When dealing with dead defs, there's a couple of edge cases. First take the
following code:

assign a, 0
assign b, 5
return a

If we simply perform a normal liveness analysis + graph coloring, we'll have
no edges between a & b (since b has no live range). As a result, there's a
chance that a and b are allocated to the same physical register, resulting
in a program that returns 5 instead of 0. One simple fix we could make here
is writing all dead defs to $zero (or running deadcode elimination before).
However, another similar edge case exists that this wouldn't easily account
for:

assign a, 10
assign r, 5
assign r, a
return r

Since the first assignment to r has no live range, r is not considered live
until the next assignment. As a result, there's no overlap between the live
ranges of a & r (when in reality there should be). Once again there's a
pretty simple fix here: we can just consider both the in-set and out-set of
a node to be simultaneously live. This is too conservative though, since it
means that a variable can never be superseded by another variable in the
same physical register.

The solution that I've found to work best here is to consider the def'd
variable live in an imaginary program point after the in set and before the
out set. Then, I create an edge from the currently def'd variable to each
member of that node's out set. In the above code, this makes r & a mutually
exclusive since a is still live in the first def to r.