ISel: Handle large (>16-bit) immediates
MIPS-32 only supports signed, 16-bit immediates in I-type instructions.
This includes all arithmetic instructions, so we need to be careful when
generating these. In the past, we just emitted all these immediates as-is
from the TigerIR input, but this would cause assembler errors when dealing
with large immediates. Instead, we now check the size of each immediate
before emitting it: if it's a valid signed-16-bit value then we can just
proceed as before; otherwise, we need to generate an additional load immediate.

Since this changed around some of the structure of our arithmetic pass, I've
gone ahead and cleaned it up a little bit. First, I defined use_op2 as a
shorthand for two use_op's. Then, I fleshed out each of the cases where we
could properly emit an I-type arithmetic instruction, resulting in quite a
bit of movement in the code and some redundant match cases disappearing. When
doing this, I noticed that we could further optimize all arithmetic with two
immediates to happen at compile-time (with the exception of division, since
a divide-by-zero should still happen at runtime when one occurs!).

Additionally, I realized that the implementation for `li` depended on this
incorrect understanding of `addi,` and went back to the drawing board.
Instead of generating an `lui` + `addi,` I opt to simply generate an `li`
pseudo-instruction. While this doesn't exist in the real MIPS hardware,
nearly all MIPS32 assemblers support it so we're okay generating it. This
resulted in some changes across the rest of the codebase but nothing major.